Rename nav.styles, split styles in nav.styles
Rename nav.styles to Nav.styles

Split off styles inside Nav.styles into:
- FiltersCard.styles
- TodoAppBar.styles
- UserCard.styles
- UserProfileMenu.styles

Update imports as needed
