+++ client/src/containers/Nav/UserProfileMenu.styles.ts
@@ -1,0 +1,14 @@
+import { makeStyles, Theme, createStyles } from '@material-ui/core/styles';
+
+
+// NOTE UserProfileMenu styling
+
+const UserProfileMenuStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    colorPrimary: {
+      color: 'lightgrey'
+    }
+  })
+)
+
+export default UserProfileMenuStyles
